+++ CHANGELOG.md
@@ -16,6 +16,7 @@
    flag to enable Mistral based judge models.
 * The `ilab` command now accepts a `-v` / `--verbose` option to enable debug logging.
   `ilab -vv` or `ilab --verbose --verbose` enables more verbose debug logging.
+* `ilab model test` generic support
 
 ### Breaking Changes